backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Cache / Queue
redis==5.2.1
celery[redis]==5.4.0

# HTTP client
httpx==0.28.1

# ML / Data
numpy==2.2.2
scipy==1.15.1
scikit-learn==1.6.1
xgboost==2.1.3
pandas==2.2.3
optuna==4.1.0

# AI Chat
google-generativeai==0.8.4

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
